loot management improvement
Added a patch on LoadLocationLoot in order to redirect the request to /coop/location/getLoot instead of /client/location/getLocalloot when connecting to a coop raid. This way, we can pass the Server Id to the server so it can provide the coop raid loot data to EFT.

Single player sessions will continue to use the default route which is /client/location/getLocalloot.

This requires the StayInTarkov.SIT.Aki-Server-Mod PR "fix-loot" to work properly.

A lot of testing was done to ensure it is working properly, but I recommend more testing to be safe.